subida del proyecto

+++ src/main/java/com/example/empresa/service/PersonaService.java
@@ -2,7 +2,10 @@ package com.example.empresa.service;
 
 
 import com.example.empresa.models.Persona;
+
 import com.example.empresa.repository.IPersonaRepository;
+import com.example.empresa.request.PersonaRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,9 +20,22 @@ public class PersonaService implements IPersonaService{
     private IPersonaRepository personaRepository;
 
 
-    public void savePersona(Persona persona){
+    public Optional<Persona> savePersona(PersonaRequest requestPersona){
 
-        personaRepository.save(persona);
+        Optional<Persona> personaAux= personaRepository.findByDni(requestPersona.getDni());
+        if(personaAux.isPresent()){
+            return Optional.empty();
+        }
+        
+
+        Persona persona= Persona.builder()
+            .nombre(requestPersona.getNombre())
+            .apellido(requestPersona.getApellido())
+            .dni(requestPersona.getDni())
+            .fechaNacimiento(requestPersona.getFechaNacimiento())
+            .build();
+
+        return Optional.of(personaRepository.save(persona));
 
 
     }
@@ -34,8 +50,14 @@ public class PersonaService implements IPersonaService{
 
 
 
-    public void deletePersonaById(Integer id) {
-        personaRepository.deleteById(id);
+    public boolean deletePersonaById(Integer id) {
+        boolean eliminado= false;
+        if(personaRepository.existsById(id)){
+            eliminado= true;
+            personaRepository.deleteById(id);    
+        }
+        
+        return eliminado;
     }
 
 
@@ -46,6 +68,21 @@ public class PersonaService implements IPersonaService{
     @Override
     public List<Persona> getPersonasMayoresDeEdad() {
         return personaRepository.getPersonasMayoresDeEdad();
+    }
+    @Override
+    public Boolean updatePersona(Integer id, PersonaRequest personaRequest) {
+        Boolean update= false;
+        Persona persona= personaRepository.getReferenceById(id);
+        if(persona != null){
+            persona.setNombre(personaRequest.getNombre());
+            persona.setApellido(personaRequest.getApellido());
+            persona.setDni(personaRequest.getDni());
+            persona.setFechaNacimiento(personaRequest.getFechaNacimiento());
+            update= true;
+            personaRepository.save(persona);
+        }
+        return update;    
+        
     }
